requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
configparser==5.0.1
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-environ==0.4.5
environ==1.0
idna==2.10
mysqlclient==2.0.2
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
